pyproject.toml
[project]
name = "mya2aprojects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.10",
    "asyncclick>=8.3.0.7",
    "baidusearch>=1.0.3",
    "bitsandbytes>=0.49.1",
    "deepagents>=0.4.4",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "google-adk>=1.17.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=3.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.1",
    "litellm>=1.79.1",
    "llama-cloud-services>=0.6.54",
    "llama-index>=0.14.12",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-dashscope>=0.5.1",
    "llama-index-llms-huggingface>=0.6.1",
    "llama-index-vector-stores-faiss>=0.5.2",
    "markdownify>=1.2.2",
    "nest-asyncio>=1.6.0",
    "psycopg2>=2.9.11",
    "sentence-transformers>=5.1.2",
    "tavily>=1.1.0",
    "timestamp>=0.0.1",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
    "tenacity>=8.2.3",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
core = [
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.3",
]
server = [
    "a2a-sdk[http-server]>=0.3.10",
    "uvicorn>=0.38.0",
    "psycopg2>=2.9.11",
]
client = [
    "asyncclick>=8.3.0.7",
    "google-adk>=1.17.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
